commit final version
angular

+++ front-end-angular-final/src/app/branchshop/branchshop-list/branchshop-list.component1.ts
@@ -1,27 +1,29 @@
-import { Component, OnInit, Input, AfterViewInit, ChangeDetectorRef, ElementRef } from '@angular/core';
-import { EmployeetypeRestApiService } from '../employeetype-rest-api.service';
+import { Component, OnInit, Input, AfterViewInit, ChangeDetectorRef } from '@angular/core';
+
 import * as $ from 'jquery';
 import 'datatables.net';
 import 'datatables.net-bs4';
 import { TokenStorageService } from 'src/app/auth/token-storage.service';
 import { Router } from '@angular/router';
-//popupstart
+import { BranchshopRestApiService } from '../branchshop-rest-api.service';
 import { MatDialog, MatDialogConfig } from "@angular/material";
-//Chọn Component làm popup
-import { EmployeetypeCreateComponent } from '../employeetype-create/employeetype-create.component';
-//popupend
+import { EmployeetypeCreateComponent } from 'src/app/employeetype/employeetype-create/employeetype-create.component';
+import { BranchshopCreateComponent } from '../branchshop-create/branchshop-create.component';
+import { EmployeeRestApiService } from 'src/app/employee/employee-rest-api.service';
+import { JsonAdaptor } from '@syncfusion/ej2-data';
 @Component({
-  selector: 'app-employeetype-list',
-  templateUrl: './employeetype-list.component.html',
-  styleUrls: ['./employeetype-list.component.css']
+  selector: 'app-branchshop-list',
+  templateUrl: './branchshop-list.component.html',
+  styleUrls: ['./branchshop-list.component.css']
 })
-export class EmployeetypeListComponent implements OnInit, AfterViewInit {
+export class BranchshopListComponent implements OnInit {
 
   //khai báo tên gọi cho 
   clients: any[];
   dataTable: any;
-  metismenu1: any;
   Content: any = [];
+  Contentcheck: any = []
+  Contentemployee: any = [];
   info: any;
   private roles: string[];
   private authority: string;
@@ -31,26 +33,17 @@ export class EmployeetypeListComponent implements OnInit, AfterViewInit {
   private authoritybrm: string;
   private authorityacc: string;
   private authoritycashier: string;
-
-  constructor(public restApi: EmployeetypeRestApiService,
+  constructor(public restApi: BranchshopRestApiService,
+    public restApiemployee: EmployeeRestApiService,
     private chRef: ChangeDetectorRef,
     private token: TokenStorageService,
-    //popupstart
-    private dialog: MatDialog,
-    //popupend
-    public router: Router, private elementRef: ElementRef) { }
+    public router: Router, private dialog: MatDialog) { }
   ngOnInit() {
     this.info = {
       token: this.token.getToken(),
       username: this.token.getUsername(),
       authorities: this.token.getAuthorities()
     };
-
-    //token start
-    if (this.token.getToken()) {
-      this.token.checklogin()
-    }
-    //token end
     if (!this.token.getToken()) {
       this.router.navigate(['login'])
     } else {
@@ -79,15 +72,12 @@ export class EmployeetypeListComponent implements OnInit, AfterViewInit {
         return false;
       });
     }
-    if (this.authorityad === 'ad' || this.authorityhr === 'hr') {
-
+    if (this.authorityacc === 'acc' || this.authoritybrm === 'brm' ||
+      this.authoritycashier === 'cashier' || this.authorityad === 'ad' || this.authorityhr === 'hr') {
     } else {
       this.router.navigate(['home'])
     }
-    this.loadEmployeetype();
-  }
-  ngAfterViewInit() {
-
+    this.loadEmployeetype()
   }
 
   loadEmployeetype() {
@@ -96,56 +86,49 @@ export class EmployeetypeListComponent implements OnInit, AfterViewInit {
       this.chRef.detectChanges();
       const table: any = $('table');
       this.dataTable = table.DataTable();
-      console.log(JSON.stringify(this.dataTable));
     })
   }
-  //popupstart
   onCreate() {
-    //token start
-    this.token.checklogin()
-    if (!this.token.getToken()) {
-      this.router.navigate(['login'])
-    } else {
-      //token end
-     
-      this.restApi.employeetypeDetails.id = null;
-      this.restApi.employeetypeDetails.name = '';
-      this.restApi.initializeFormGroup();
-      const dialogConfig = new MatDialogConfig();
-      dialogConfig.disableClose = true;
-      dialogConfig.autoFocus = true;
-      dialogConfig.width = "60%";
-      //Chọn Component làm popup
-      this.dialog.open(EmployeetypeCreateComponent, dialogConfig);
-    }  // thêm dấu ngoặc
+    this.restApi.employeetypeDetails.id = null;
+    this.restApi.employeetypeDetails.address = '';
+    this.restApi.employeetypeDetails.name = '';
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    dialogConfig.width = "60%";
+    //Chọn Component làm popup
+    this.dialog.open(BranchshopCreateComponent, dialogConfig);
   }
   onUpdate(employeetype) {
-    this.token.checklogin()
-    if (!this.token.getToken()) {
-      this.router.navigate(['login'])
-    } else {
     this.restApi.employeetypeDetails = employeetype;
-    this.restApi.editFormGroup(employeetype);
     const dialogConfig = new MatDialogConfig();
     dialogConfig.disableClose = true;
     dialogConfig.autoFocus = true;
     dialogConfig.width = "60%";
     //Chọn Component làm popup dùng chung create cho update
-    this.dialog.open(EmployeetypeCreateComponent, dialogConfig);
-  }}
-  //popupend
+    this.dialog.open(BranchshopCreateComponent, dialogConfig);
+  }
   // Delete employee
   deleteEmployeetype(id) {
-    this.token.checklogin()
-    if (!this.token.getToken()) {
-      this.router.navigate(['login'])
-    } else {
-    if (window.confirm('Are you sure, you want to delete?')) {
-      this.restApi.deleteEmployeetype(id).subscribe(data => {
-        this.loadEmployeetype()
-      })
-    }
-  }}
+    this.restApiemployee.getEmployeebybranchshopid(id).subscribe((data: {}) => {
+      this.Contentemployee = data;
+      this.Contentcheck = this.Contentemployee.content;
+      console.log('so nhan vien: ' + (this.Contentcheck.length))
+      if (this.Contentcheck.length < 1) {
+        if (window.confirm('Are you sure, you want to delete?')) {
+          this.restApi.deleteEmployeetype(id).subscribe(data => {
+            this.loadEmployeetype()
+          })
+        }
+      } else {
+        window.confirm('Chi nhánh hiện còn nhân viên bạn không thể xóa')
+        // window.alert(this.Contentcheck.length);
+        // console.log(this.Contentcheck.length)
+      }
+    })
+
+
+  }
   logout() {
     this.token.signOut();
     this.router.navigate(['login'])
@@ -153,3 +136,4 @@ export class EmployeetypeListComponent implements OnInit, AfterViewInit {
   }
 
 }
+
